requirements.txt
selenium
pytest
webdriver-manager
pytest-html
python-dotenv
opencv-python
numpy
pyautogui
Pillow
